Database schema overhaul.
Removed redundant fields and converted reviews into their own collection.
