--- src/main/java/org/suitesquad/likehome/model/Address.java
@@ -1,9 +1,0 @@
-package org.suitesquad.likehome.model;
-
-public class Address {
-    private String streetAddress;
-    private String city;
-    private String state;
-    private String zipCode;
-    private String country;
-}
