--- src/main/java/org/suitesquad/likehome/model/ContactInfo.java
@@ -1,7 +1,0 @@
-package org.suitesquad.likehome.model;
-
-public class ContactInfo {
-    private String phone;
-    private String email;
-    private String website;
-}
