+++ src/main/java/org/suitesquad/likehome/model/Reservation.java
@@ -17,7 +17,7 @@ public class Reservation {
     @Indexed(unique = true)
     private String id;
 
-    //@Indexed
+    @Indexed
     private String userId;
 
     //@Indexed
