--- src/main/java/org/suitesquad/likehome/model/Room.java
@@ -1,11 +1,0 @@
-package org.suitesquad.likehome.model;
-
-import java.util.List;
-
-public class Room {
-    private String roomType;
-    private double price;   // price per night
-    private int availability;
-    private List<String> features;
-    private String imageUrl;
-}
